[OPEN-362] Hold access tokens in memory after refresh (#4)
This PR fixes issues with unexpected logout states caused by the
unreliability of propagation of `document.cookie` updates within a hook.

We still rely on `document.cookie` for initial load, but we hold the
`accessToken` returned from the `refresh` call in memory in order to
prevent false negatives from outdated `document.cookie` values.

This has been tested in a production app.