Suspend on expired access tokens (#13)
In the case where a user has a working refresh token, and visits a
TesseralProvider-protected page with an expired access token,
TesseralProvider today optimistically renders with that expired access
token, and immediately begins refreshing it.

This PR changes that behavior. Expired access tokens cause a suspended
state. The logic for triggering pre-refreshing is unchanged.

This change is to reduce a class of errors for users, where the expired
access tokens make fetches that fail. It's very common for that to have
negative effects in terms of faulty redirects or other irrecoverable
error states on the frontend. It also causes noise in bugsnaggers.

In service of making the pre-refreshing more accurate, this PR also
updates the checking period for these to 2 seconds. This PR makes our
behavior change in the 10-second window at the end of an access token's
lifetime. This has a sort of Nyquist rate effect, and making the
(inexpensive) expiry checks happen at least twice as often as the
10-second period makes this code more obviously correct.